requirements.txt
Django==4.1.3
djangorestframework
djangorestframework-simplejwt
django-cors-headers
drf-yasg

